two unit test still fail in odd places. 32->4096 and 2048->4096

+++ cpp/Image.cuh
@@ -105,7 +105,7 @@ float ReturnSumOfReal(float* input, short4 size)
     {
       for (int i = 0; i < size.x; i++)
       {
-        temp_sum += (input[address]);
+        temp_sum += double(input[address]);
         address++;
       }
       address += padding_jump_val;
